modular-base-agent: add support for extending the base agent with modules
adding new module is done using the `libs` directory when a user
installs a new plugin he has to create a `.js` file at libs folder.
the file must contain the name of the module (unique name) and the
function as described at the config.js file given to the server,
each method name is a function.