register-agent: support registering agent with key
when registering agent the server saves it's key.
when action is executed the server sends the key to the relevant agent
and the agent execute the command onlly if the key is correct.
key is 128 bit of hex value.
registering is done using
`node utils/register.js agentUrl username passwd serverUrl`